+++ src/lib/components/admin/PaymentsTab.svelte
@@ -1,5 +1,6 @@
 <script>
     import { onMount } from "svelte";
+    import { auth } from "../../stores/auth";
     import { api } from "../../api.svelte";
     import {
         Search,
@@ -194,15 +195,17 @@
                                     >
                                         <Edit2 size={18} />
                                     </button>
-                                    <button
-                                        on:click={() =>
-                                            handleDelete(payment._id)}
-                                        class="p-2 text-red-600 hover:bg-red-50 rounded-lg transition-colors"
-                                        title="Delete"
-                                        aria-label="Delete payment"
-                                    >
-                                        <Trash2 size={18} />
-                                    </button>
+                                    {#if $auth.isAdmin}
+                                        <button
+                                            on:click={() =>
+                                                handleDelete(payment._id)}
+                                            class="p-2 text-red-600 hover:bg-red-50 rounded-lg transition-colors"
+                                            title="Delete"
+                                            aria-label="Delete payment"
+                                        >
+                                            <Trash2 size={18} />
+                                        </button>
+                                    {/if}
                                 </div>
                             </td>
                         </tr>
